requirements.txt
amqp==5.3.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Events==0.5
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.0
MarkupSafe==3.0.2
opensearch-py==2.8.0
packaging==24.2
prompt_toolkit==3.0.50
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
